+++ tests/slave_test.rs
@@ -359,7 +359,11 @@ async fn test_fc17_report_server_id() {
     let (slave, server, _coils, _di, _hr, _ir) = create_slave().await;
     let master = create_master(&server).await;
 
-    let res = master.report_server_id(UNIT, 1, None).await.unwrap().unwrap();
+    let res = master
+        .report_server_id(UNIT, 1, None)
+        .await
+        .unwrap()
+        .unwrap();
     assert_eq!(res.server_id, vec![UNIT]);
     assert!(res.run_indicator_status);
     assert_eq!(res.additional_data, vec![1, 2, 3]);